Fetch APM versions for variants
