+++ src/tasks/publish/types.ts
@@ -1,4 +1,6 @@
+import { Manifest } from "@dappnode/types";
 import { VerbosityOptions } from "../../commands/build/types.js";
+import { ManifestFormat } from "../../files/manifest/types.js";
 import { UploadTo } from "../../releaseUploader/index.js";
 import { ReleaseType } from "../../types.js";
 
@@ -15,6 +17,13 @@ export interface PublishOptions {
   developerAddress?: string;
   githubRelease?: boolean;
   verbosityOptions: VerbosityOptions;
-  variantsDirPath?: string;
-  variants?: string[];
+  variantsDirPath: string;
+  variants: string[] | null;
+}
+
+export interface ManifestsMap {
+  [variant: string]: {
+    manifest: Manifest;
+    format: ManifestFormat;
+  };
 }
